move SubprocessGit to server_test.go
