+++ enterprise/internal/rockskip/server_test.go
@@ -1,12 +1,15 @@
 package rockskip
 
 import (
+	"bufio"
 	"context"
 	"fmt"
+	"io"
 	"os"
 	"os/exec"
 	"path"
 	"sort"
+	"strconv"
 	"strings"
 	"testing"
 
@@ -174,4 +177,133 @@ func TestIndex(t *testing.T) {
 
 	rm("a.txt")
 	commit("rm a.txt")
+}
+
+type SubprocessGit struct {
+	gitDir        string
+	catFileCmd    *exec.Cmd
+	catFileStdin  io.WriteCloser
+	catFileStdout bufio.Reader
+}
+
+func NewSubprocessGit(gitDir string) (*SubprocessGit, error) {
+	cmd := exec.Command("git", "cat-file", "--batch")
+	cmd.Dir = gitDir
+
+	stdin, err := cmd.StdinPipe()
+	if err != nil {
+		return nil, err
+	}
+
+	stdout, err := cmd.StdoutPipe()
+	if err != nil {
+		return nil, err
+	}
+
+	err = cmd.Start()
+	if err != nil {
+		return nil, err
+	}
+
+	return &SubprocessGit{
+		gitDir:        gitDir,
+		catFileCmd:    cmd,
+		catFileStdin:  stdin,
+		catFileStdout: *bufio.NewReader(stdout),
+	}, nil
+}
+
+func (git SubprocessGit) Close() error {
+	err := git.catFileStdin.Close()
+	if err != nil {
+		return err
+	}
+	return git.catFileCmd.Wait()
+}
+
+func (git SubprocessGit) LogReverseEach(repo string, givenCommit string, n int, onLogEntry func(entry LogEntry) error) (returnError error) {
+	log := exec.Command("git", LogReverseArgs(n, givenCommit)...)
+	log.Dir = git.gitDir
+	output, err := log.StdoutPipe()
+	if err != nil {
+		return err
+	}
+
+	err = log.Start()
+	if err != nil {
+		return err
+	}
+	defer func() {
+		err = log.Wait()
+		if err != nil {
+			returnError = err
+		}
+	}()
+
+	return ParseLogReverseEach(output, onLogEntry)
+}
+
+func (git SubprocessGit) RevListEach(repo string, givenCommit string, onCommit func(commit string) (shouldContinue bool, err error)) (returnError error) {
+	revList := exec.Command("git", RevListArgs(givenCommit)...)
+	revList.Dir = git.gitDir
+	output, err := revList.StdoutPipe()
+	if err != nil {
+		return err
+	}
+
+	err = revList.Start()
+	if err != nil {
+		return err
+	}
+	defer func() {
+		err = revList.Wait()
+		if err != nil {
+			returnError = err
+		}
+	}()
+
+	return RevListEach(output, onCommit)
+}
+
+func (git SubprocessGit) ArchiveEach(repo string, commit string, paths []string, onFile func(path string, contents []byte) error) error {
+	for _, path := range paths {
+		_, err := git.catFileStdin.Write([]byte(fmt.Sprintf("%s:%s\n", commit, path)))
+		if err != nil {
+			return errors.Wrap(err, "writing to cat-file stdin")
+		}
+
+		line, err := git.catFileStdout.ReadString('\n')
+		if err != nil {
+			return errors.Wrap(err, "read newline")
+		}
+		line = line[:len(line)-1] // Drop the trailing newline
+		parts := strings.Split(line, " ")
+		if len(parts) != 3 {
+			return errors.Newf("unexpected cat-file output: %q", line)
+		}
+		size, err := strconv.ParseInt(parts[2], 10, 64)
+		if err != nil {
+			return errors.Wrap(err, "parse size")
+		}
+
+		fileContents, err := io.ReadAll(io.LimitReader(&git.catFileStdout, size))
+		if err != nil {
+			return errors.Wrap(err, "read contents")
+		}
+
+		discarded, err := git.catFileStdout.Discard(1) // Discard the trailing newline
+		if err != nil {
+			return errors.Wrap(err, "discard newline")
+		}
+		if discarded != 1 {
+			return errors.Newf("expected to discard 1 byte, but discarded %d", discarded)
+		}
+
+		err = onFile(path, fileContents)
+		if err != nil {
+			return errors.Wrap(err, "onFile")
+		}
+	}
+
+	return nil
 }
